+++ lib/src/components/badge/badge_style.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:impaktfull_ui/src/models/asset.dart';
 import 'package:impaktfull_ui/src/theme/theme.dart';
+import 'package:impaktfull_ui/src/util/extension/color_extensions.dart';
 
 class ImpaktfullUiBadgeTheme extends ImpaktfullUiComponentTheme {
   final ImpaktfullUiBadgeAssetsTheme assets;
@@ -17,6 +18,49 @@ class ImpaktfullUiBadgeTheme extends ImpaktfullUiComponentTheme {
 
   static ImpaktfullUiBadgeTheme of(BuildContext context) =>
       ImpaktfullUiTheme.of(context).components.badge;
+
+  static ImpaktfullUiBadgeTheme getDefault({
+    required ImpaktfullUiAssetTheme assets,
+    required ImpaktfullUiColorTheme colors,
+    required ImpaktfullUiTextStylesTheme textStyles,
+    required ImpaktfullUiDimensTheme dimens,
+    required ImpaktfullUiDurationTheme durations,
+    required ImpaktfullUiShadowsTheme shadows,
+  }) =>
+      ImpaktfullUiBadgeTheme(
+        assets: ImpaktfullUiBadgeAssetsTheme(
+          close: assets.icons.close,
+        ),
+        colors: ImpaktfullUiBadgeColorTheme(
+          primaryText: colors.accent,
+          primaryBorder: colors.accent,
+          primaryBackground: colors.accent.withOpacityPercentage(0.05),
+          secondaryText: colors.secondary,
+          secondaryBorder: colors.secondary.withOpacityPercentage(0.5),
+          secondaryBackground: colors.secondary.withOpacityPercentage(0.05),
+          neutralText: colors.text,
+          neutralBorder: colors.border,
+          neutralBackground: colors.border.withOpacityPercentage(0.05),
+          errorText: colors.error,
+          errorBorder: colors.error.withOpacityPercentage(0.5),
+          errorBackground: colors.error.withOpacityPercentage(0.05),
+          warningText: colors.warning,
+          warningBorder: colors.warning.withOpacityPercentage(0.5),
+          warningBackground: colors.warning.withOpacityPercentage(0.05),
+          successText: colors.success,
+          successBorder: colors.success.withOpacityPercentage(0.5),
+          successBackground: colors.success.withOpacityPercentage(0.05),
+        ),
+        dimens: ImpaktfullUiBadgeDimensTheme(
+          borderRadius: dimens.borderRadiusCircle,
+          borderWidth: 1,
+        ),
+        textStyles: ImpaktfullUiBadgeTextStylesTheme(
+          small: textStyles.onCard.text.small,
+          medium: textStyles.onCard.text.small,
+          large: textStyles.onCard.text.small,
+        ),
+      );
 }
 
 class ImpaktfullUiBadgeAssetsTheme {
@@ -71,9 +115,11 @@ class ImpaktfullUiBadgeColorTheme {
 
 class ImpaktfullUiBadgeDimensTheme {
   final BorderRadiusGeometry borderRadius;
+  final double borderWidth;
 
   const ImpaktfullUiBadgeDimensTheme({
     required this.borderRadius,
+    required this.borderWidth,
   });
 }